requirements.txt
requests
pandas
openpyxl
python-dotenv
google-generativeai
streamlit
plotly
matplotlib
